// api/api.ts
export const AuthMethod = {
    Password: 'Password',
    Email: 'Email',
    Mobile: 'Mobile'
} as const;

export type AuthMethod = typeof AuthMethod[keyof typeof AuthMethod];
// ...
export interface LoginHistoryDto {
    /**
     * 
     * @type {string}
     * @memberof LoginHistoryDto
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LoginHistoryDto
     */
    'email'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LoginHistoryDto
     */
    'location'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LoginHistoryDto
     */
    'timestamp'?: string;
    /**
     * 
     * @type {AuthMethod}
     * @memberof LoginHistoryDto
     */
    'method'?: AuthMethod;
    /**
     * 
     * @type {string}
     * @memberof LoginHistoryDto
     */
    'platform'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LoginHistoryDto
     */
    'userAgent'?: string | null;
}
/**
 * 
 * @export
 * @enum {string}
 */

export const LoginHistoryFilter = {
    All: 'All',
    OnlyMe: 'OnlyMe'
} as const;

export type LoginHistoryFilter = typeof LoginHistoryFilter[keyof typeof LoginHistoryFilter];
// ...
export interface ServiceDto {
    /**
     * 
     * @type {string}
     * @memberof ServiceDto
     */
    'id'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceDto
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceDto
     */
    'description'?: string | null;
    /**
     * 
     * @type {AuthMethod}
     * @memberof ServiceDto
     */
    'authMethod'?: AuthMethod;
    /**
     * 
     * @type {ServiceType}
     * @memberof ServiceDto
     */
    'serviceType'?: ServiceType;
    /**
     * 
     * @type {ServiceState}
     * @memberof ServiceDto
     */
    'serviceState'?: ServiceState;
    /**
     * 
     * @type {string}
     * @memberof ServiceDto
     */
    'organization'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceDto
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceDto
     */
    'url'?: string | null;
}
// ...
export const ServiceState = {
    Registered: 'REGISTERED',
    Pending: 'PENDING'
} as const;

export type ServiceState = typeof ServiceState[keyof typeof ServiceState];


/**
 * 
 * @export
 * @enum {string}
 */

export const ServiceType = {
    Website: 'Website',
    Mobile: 'Mobile',
    Desktop: 'Desktop'
} as const;

export type ServiceType = typeof ServiceType[keyof typeof ServiceType];
// ...
export class ServicesApi extends BaseAPI {
    /**
     * 
     * @param {string} [serviceId] 
     * @param {LoginHistoryFilter} [filter] 
     * @param {number} [pageNumber] 
     * @param {number} [pageSize] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ServicesApi
     */
    public servicesGetLoginHistory(serviceId?: string, filter?: LoginHistoryFilter, pageNumber?: number, pageSize?: number, options?: AxiosRequestConfig) {
        return ServicesApiFp(this.configuration).servicesGetLoginHistory(serviceId, filter, pageNumber, pageSize, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} serviceId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ServicesApi
     */
    public servicesGetServiceDetail(serviceId: string, options?: AxiosRequestConfig) {
        return ServicesApiFp(this.configuration).servicesGetServiceDetail(serviceId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} [searchQuery] 
     * @param {number} [pageNumber] 
     * @param {number} [pageSize] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ServicesApi
     */
    public servicesGetServices(searchQuery?: string, pageNumber?: number, pageSize?: number, options?: AxiosRequestConfig) {
        return ServicesApiFp(this.configuration).servicesGetServices(searchQuery, pageNumber, pageSize, options).then((request) => request(this.axios, this.basePath));
    }
}